backend/requirements.txt
pydantic>=2.0.0
pydantic-settings
numpy
pandas
pytest
mypy
ruff
sqlalchemy>=2.0.0
sqlmodel
qdrant-client
requests
python-dotenv
yfinance
uuid
fastapi
uvicorn
httpx
pyyaml
psutil
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
